// backend/src/utils/jwt.d.ts
export declare const generateTokens: (userId: string) => {
    accessToken: never;
    refreshToken: never;
};
export declare const verifyAccessToken: (token: string) => {
    userId: string;
};
export declare const verifyRefreshToken: (token: string) => {
    userId: string;
};
//# sourceMappingURL=jwt.d.ts.map